requirements.txt
strands-agents
bedrock-agentcore
google-play-scraper
#app-store-scraper
#boto3
